File Change isAuthenticate Action
