+++ conf/sub.api.routes
@@ -3,7 +3,10 @@
 # https://www.playframework.com/documentation/latest/ScalaRouting
 # ~~~~
 
+# authenticate API
 POST   /auth/login            controllers.api.authenticate.UserLoginController.login
 POST   /auth/signup           controllers.api.authenticate.UserSignupController.signup
 DELETE /auth/logout           controllers.api.authenticate.UserLogoutController.logout
-GET    /auth/isAuthenticate   controllers.api.authenticate.UserSignupController.isAuthenticate
+
+# validation API
+GET    /auth/isAuthenticate   controllers.api.service.ValidationController.isAuthenticate
